refactor module name
common is a public module
